pyproject.toml
[project]
name = "langgraph-chatbot"
version = "0.1.0"
description = "Simple LangGraph CLI chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph",
  "langchain",
  "langchain-openai",
  "python-dotenv",
  "fastapi[standard]",
  "requests",
  "gradio>=4.0.0"
]

[project.scripts]
chatbot = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false


